requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.31
pydantic==2.8.2
pydantic[email]==2.8.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
Jinja2==3.1.4
Pillow==10.4.0
python-barcode==0.15.1
pandas==2.2.2
openpyxl==3.1.5
starlette==0.37.2
